Breaks dirs/files off into own files
