+++ generators/app/projectDirectories.js
@@ -1,0 +1,13 @@
+//  The directories for the app
+const projectDirectories = [
+  './public/',
+  './server/',
+  './src/',
+  './src/actions/',
+  './src/components/',
+  './src/css/',
+  './src/images/',
+  './src/reducers/'
+];
+
+module.exports = projectDirectories;
