+++ generators/app/templates/_package.json
@@ -32,6 +32,7 @@
     "babel-core": "^6.9.0",
     "babel-eslint": "^6.0.4",
     "babel-loader": "^6.2.4",
+    "babel-plugin-remove-console": "^1.0.1",
     "babel-plugin-transform-object-assign": "^6.8.0",
     "babel-plugin-transform-react-remove-prop-types": "^0.2.7",
     "babel-preset-es2015": "^6.9.0",
